feat: detector without float or div

+++ lib/detector/Detector.h
@@ -7,13 +7,13 @@ class Detector {
 private:
     int trigPin;
     int echoPin;
-    static constexpr float SOUND_SPEED = 340.0;  // Vitesse du son en m/s
+    static constexpr int SOUND_SPEED = 340;  // Vitesse du son en m/s
     static constexpr int TRIG_PULSE_DURATION_US = 10;
 
 public:
     Detector(int trig, int echo);
     void begin();
-    float getDistance();
+    int getDistance();
 };
 
 #endif
